+++ Assets/Scripts/GamePlay/HumbleObjects/Interfaces/ICameraHumbleObject.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 namespace Source.GamePlay.Controllers.Interfaces
 {
-    public interface ICameraController
+    public interface ICameraHumbleObject
     {
         Ray? CameraScreenPointToRay(Vector2 mousePosition);
         void RigidBodySetDamping(float dampingForce);
