--- Assets/Scripts/GamePlay/Services/Units/Unit.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 12dd5ffc401e6b04fb6d71d2d9bac56b
