--- Assets/Scripts/GamePlay/Services/Units/UnitService.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: e81566f51c001ad4f956f7af3e55ec53
