+++ Assets/Scripts/Shared/HumbleObjects/Interfaces.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e2e856b9336c1ba428c88e0a74cf640f
+guid: f9e7076c7874697459362dc857b651b9
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
